% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{ADMM}
\alias{ADMM}
\title{The alternating direction method of multipliers (ADMM) algorithm for subgroup analysis with the SCAD penalty}
\usage{
ADMM(x, y, lambda, n, theta, gamma, epsilon)
}
\arguments{
\item{x}{n*p design matrix with p predictors}

\item{y}{n-dimensional response vector}

\item{lambda}{the penalty parameter of SCAD}

\item{n}{the sample size}

\item{theta}{another penalty parameter}

\item{gamma}{a parameter that controls the concavity of SCAD}

\item{epsilon}{the iteration accuracy}
}
\value{
the estimators of intercept, cofficient and the result of subgroup
}
\description{
The ADMM algorithm implements subgroup analysis based on regression models
}
\examples{
\dontrun{
n <- 100
p <- 5
x <- matrix(0,n,p)
sigma <- matrix(c(rep(0.3,p^2)), p, p)
for(i in 1:5) sigma[i,i] <- 1
x <- mvrnorm(n,rep(0,p),sigma,empirical = TRUE)
eps <- as.matrix(rnorm(n,0,0.25)) 
beta0 <- matrix(0,1,p)
beta0 <- runif(p,0.5,1)
u0 <- as.matrix(2*rbinom(n,1,0.5)-1)   
y0 <- matrix(0,n,1)
for(i in 1:n) y0[i,] <- u0[i,1]+t(x[i,])\%*\%beta0+eps[i,]
ADMM(x, y0, 0.45, 100, 1, 3, 1e-5)
}
}
