% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CMetropolis}
\alias{CMetropolis}
\title{A random number generator using Rcpp}
\usage{
CMetropolis(sigma = 0.05, x0 = 25, N = 5000L)
}
\arguments{
\item{sigma}{the standard deviation of normal distribution}

\item{x0}{the initial value}

\item{N}{the number of samples}
}
\value{
a chain of random numbers
}
\description{
A random number generator using Rcpp
}
\examples{
\dontrun{
CMetropolis(0.05,25,5000)
}
}
