% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{mean_imputation}
\alias{mean_imputation}
\title{The mean imputation method for modifying right censoring survival time data}
\usage{
mean_imputation(y0, n, per)
}
\arguments{
\item{y0}{the n-dimensional vector for the logarithm of failure time}

\item{n}{the sample size}

\item{per}{the censoring rate}
}
\value{
the modified survival data
}
\description{
The mean imputation method for modifying right censoring survival time data with K-M estimator
}
\examples{
\dontrun{
n <- 100
p <- 5
x <- matrix(0,n,p)
sigma <- matrix(c(rep(0.3,p^2)), p, p)
for(i in 1:5) sigma[i,i] <- 1
x <- mvrnorm(n,rep(0,p),sigma,empirical = TRUE)
eps <- as.matrix(rnorm(n,0,0.25)) 
beta0 <- matrix(0,1,p)
beta0 <- runif(p,0.5,1)
u0 <- as.matrix(2*rbinom(n,1,0.5)-1)   
y0 <- matrix(0,n,1)
for(i in 1:n) y0[i,] <- u0[i,1]+t(x[i,])\%*\%beta0+eps[i,]
mean_imputation(y0, 100, 0.1)
}
}
