% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{parking_solution}
\alias{parking_solution}
\title{A method for solving an interesting parking problem}
\usage{
parking_solution(x, n)
}
\arguments{
\item{x}{the length of parking interval}

\item{n}{numbers of independent repeated trials}
}
\value{
numbers of parking cars and the average of parking cars in unit length
}
\description{
This method can solve the parking problem in fixed interval
}
\examples{
\dontrun{
parking_solution(1000, 1000)
}
}
