\name{StatComp20005-package}
\alias{StatComp20005-package}
\alias{StatComp20005}
\docType{package}
\title{
 This R package presents the homework of statistic computing course by author 20005.
}
\description{
  The R package contains three functions written by author 20005 and all homework of the course called statistic computing
}
\details{
The R package includes three functions. Two of them are relevant to the undergraduate thesis of author. Another one is a function that can be used to solve an interesting parking problem.
}
\author{
Zhaohui Xu

Maintainer: Zhaohui Xu <xzh2517@mail.ustc.edu.cn>
}
\references{
 A concave pairwise fusion approach to subgroup analysis.
 Predicting patient survival from microarray data by accelerated failure time modeling using
partial least squares and LASSO.
}
\keyword{ package }
